feat(teams): add TanStack Query client component to team list page
Co-locates teamKeys factory and moves rendering logic to TeamsListClient,
hydrating from SSR initialData to avoid duplicate fetching.
